Moves MarginsConverter to System.Windows.Extensions
