+++ src/System.Drawing.Common/src/System/Drawing/Printing/Margins.cs
@@ -10,6 +10,9 @@ namespace System.Drawing.Printing
     /// <summary>
     /// Specifies the margins of a printed page.
     /// </summary>
+#if netcoreapp
+    [TypeConverter("System.Drawing.Printing.MarginsConverter, System.Windows.Extensions, Version=4.0.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51")]
+#endif
     public partial class Margins : ICloneable
     {
         private int _left;
