+++ src/System.Windows.Extensions/ref/System.Windows.Extensions.cs
@@ -62,7 +62,19 @@ namespace System.Drawing
         public override bool GetStandardValuesSupported(System.ComponentModel.ITypeDescriptorContext context) { throw null; }
     }
 }
-
+namespace System.Drawing.Printing
+{
+    public partial class MarginsConverter : System.ComponentModel.ExpandableObjectConverter
+    {
+        public MarginsConverter() { }
+        public override bool CanConvertFrom(System.ComponentModel.ITypeDescriptorContext context, System.Type sourceType) { throw null; }
+        public override bool CanConvertTo(System.ComponentModel.ITypeDescriptorContext context, System.Type destinationType) { throw null; }
+        public override object ConvertFrom(System.ComponentModel.ITypeDescriptorContext context, System.Globalization.CultureInfo culture, object value) { throw null; }
+        public override object ConvertTo(System.ComponentModel.ITypeDescriptorContext context, System.Globalization.CultureInfo culture, object value, System.Type destinationType) { throw null; }
+        public override object CreateInstance(System.ComponentModel.ITypeDescriptorContext context, System.Collections.IDictionary propertyValues) { throw null; }
+        public override bool GetCreateInstanceSupported(System.ComponentModel.ITypeDescriptorContext context) { throw null; }
+    }
+}
 namespace System.Security.Cryptography.X509Certificates
 {
     public enum X509SelectionFlag
